Rename tizen-error to tizen-error-container and add line at end of error.less
